refactor: Improved name for user union permissions module. Also added fields roles and permissions to UserUnionPermission entity. Added alias for enums. Removed permission entity.
