+++ package.json
@@ -78,7 +78,8 @@
       "^@/(.*)$": "<rootDir>/../src/$1",
       "^@helpers/(.*)$": "<rootDir>/../src/common/helpers/$1",
       "^@models/(.*)$": "<rootDir>/../src/models/$1",
-      "^@our-types/(.*)$": "<rootDir>/../src/common/types/$1"
+      "^@our-types/(.*)$": "<rootDir>/../src/common/types/$1",
+      "^@enums/(.*)$": "<rootDir>/../src/common/enums/$1"
     }
   }
 }
