+++ src/common/enums/permission.enum.ts
@@ -1,0 +1,1 @@
+export enum Permission {}
