+++ src/common/enums/role.enum.ts
@@ -1,0 +1,1 @@
+export enum Role {}
