--- src/models/permissions/permissions.controller.ts
@@ -1,4 +1,0 @@
-import { Controller } from '@nestjs/common';
-
-@Controller('permissions')
-export class PermissionsController {}
