--- src/models/permissions/permissions.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PermissionsService } from './permissions.service';
-import { PermissionsController } from './permissions.controller';
-
-@Module({
-  providers: [PermissionsService],
-  controllers: [PermissionsController],
-})
-export class PermissionsModule {}
