+++ src/models/user-union-permissions/user-union-permissions.module.ts
@@ -1,5 +1,5 @@
 import { Module } from '@nestjs/common';
-import { UserUnionPermissionsController } from './user_union_permissions.controller';
+import { UserUnionPermissionsController } from './user-union-permissions.controller';
 
 @Module({
   controllers: [UserUnionPermissionsController],
