+++ test/jest-e2e.json
@@ -10,6 +10,7 @@
     "^@/(.*)$": "<rootDir>/../src/$1",
     "^@helpers/(.*)$": "<rootDir>/../src/common/helpers/$1",
     "^@models/(.*)$": "<rootDir>/../src/models/$1",
-    "^@our-types/(.*)$": "<rootDir>/../src/common/types/$1"
+    "^@our-types/(.*)$": "<rootDir>/../src/common/types/$1",
+    "^@enums/(.*)$": "<rootDir>/../src/common/enums/$1"
   }
 }
